ShortcutHelper takes in an Intent action
Set the package name in the Intent so that they are only received by Chrome.
    
This has the unfortunate side-effect of shortcuts remaining on screen after Chrome is uninstalled, but it allows us to redirect where Intents are handled instead of forcing them to be handled forever by the named class.
    
BUG=282640
TEST=ShortcutHelperTest

Review URL: https://chromiumcodereview.appspot.com/23842008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@222997 0039d316-1c4b-4281-b951-d872f2087c98