Fix PrintingControllerTest#testNormalPrintingFlow and reenable the test
Moves the creation of PrintingControllerImpl into the UI thread.
Before, it wasn't being created in the UI thread, and for some reason
PrintingControllerImpl#create wasn't failing its UI thread assertion.
It did start to fail for N4 with KitKat, so this fixes the issue and
re-enables the test.

BUG=330291

Review URL: https://codereview.chromium.org/123883002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@243850 0039d316-1c4b-4281-b951-d872f2087c98