Fix a new fullscreen video crash if ContentVideoView cannot be created.
Test url: http://m.discovery.com/tv-shows/dual-survival/videos/change-is-coming-to-dual-survival

Sometimes clicking on a video will open up a new tab. In this case, the old tab will lose its fullscreen manager.
If the video also tries to enter fullscreen on the same click, then onShowCustomView() will fail to find the fullscreen manager.
We should let the video exit fullscreen in this case, this matches the behavior of that of the old fullscreen video.

BUG=366192

Review URL: https://codereview.chromium.org/250853005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267204 0039d316-1c4b-4281-b951-d872f2087c98