[App banners] Break apart AppBannerManager and AppBannerInfoBarDelegate
Moves a bunch of code from AppBannerManager into the new Java-side
AppBannerInfoBarDelegate.  This effectively makes it so that once
an app banner is created by the AppBannerManager the AppBannerInfoBar
is on its own for managing installation and blocking of the app it's
promoting.

Main changes include getting the WebContents and Tab from the infobar()
itself rather than relying on the AppBannerManager to provide it.

Further cleanup is required on the AppBannerManager to clean up the
creation pipline.

BUG=457414

Review URL: https://codereview.chromium.org/921073002

Cr-Commit-Position: refs/heads/master@{#316345}