Android Sync test for downloading typed URLs
This CL adds an instrumentation test case, testDownloadTypedUrl, that
does the following:
1) Signs in to Sync,
2) Verifies no typed URLs exist on the client,
3) Injects a typed URL into the fake Sync server, and
4) Verifies that the typed URL is downloaded from the server.

Steps 2 and 4 are currently performed in a less-than-ideal fashion
(local entity counts are checked, but actual data is not).

Step 3 requires infrastructure that is included in this CL. A follow-up
CL is required to generalize the injection logic for all data types.

BUG=365774

Review URL: https://codereview.chromium.org/948463004

Cr-Commit-Position: refs/heads/master@{#320032}