Add Robolectric unittest suites to all.gyp and rename build targets.
Renaming chrome_shell_unit_tests to chrome_junit_tests as it tests code in the directory of the chrome_java target (and not the chrome_shell_java target). Same thing for content_shell_unit_tests. And adding chrome_junit_tests and content_junit_tests to the all.gyp file so we can run the tests on some of the bots.

BUG=

Review URL: https://codereview.chromium.org/1039733002

Cr-Commit-Position: refs/heads/master@{#322813}