Fix chrome_shell_java deps
The native parts of dom_distiller will try to register jni methods on
startup. This fails if the apk doesn't contain the corresponding java
bits. So, even though chrome_shell_apk doesn't actually use the java
stuff in dom_distiller_content_java, it still needs to depend on it and
include it in the apk.

Review URL: https://codereview.chromium.org/1090153007

Cr-Commit-Position: refs/heads/master@{#326937}