New UMA settings fragment using Android prefs.
Previous reverted CL is here (https://codereview.chromium.org/1080723002/) and the first patch set also corresponds to it.

The problem for revert was in profile_io_data.cc file because of initiating two preferences for Android. The kMetricsReportingEnabled was introduced for future consolidation of two preferences. I have removed the initialization for kMetricsReportingEnabled which shouldn't change the behavior as the metrics service is still using kCrashReportingEnabled and Android side updates both preferences.

Rest of the code is not changed and corresponds to the version that was LGTM-ed before.

BUG=455847

TBR=dfalcantara@chromium.org,nyquist@chromium.org

Review URL: https://codereview.chromium.org/1058433007

Cr-Commit-Position: refs/heads/master@{#327047}