Change BaseHttpTestServer to use callbacks.
As part of adding support for adding new data for feedback reports
on Android, there is a need to ensure that some responses are delayed
on the server.

This already works today by adding a call to sleep() on the server
thread. However, this does not work when there are several
simulataneous requests and not all of their responses should be
delayed as the first request that triggers sleep() would block all
subsequent requests.

This CL adds support in the BaseHttpTestServer and the subclasses to
handle all responses by passing around an HttpResponseCallback,
instead of requiring handlers to return the response synchronously.

The ConnectivityCheckerTest.ConnectivityTestServer takes advantage of
this to create an internal HandlerThread when it starts up and use
the associated Handler to post delayed tasks to whenever a delay
is required for a test.

BUG=386395

Review URL: https://codereview.chromium.org/1129153006

Cr-Commit-Position: refs/heads/master@{#331685}