Always open bookmarks in incognito mode in the foreground
This launch flag has been incorrect since forever ago, but its being incorrect
affects only document mode because it expects the flag to have a sane value.
Change it so that incognito tabs opened for bookmarks are opened in incognito
mode and add a unit test to make sure that incorrect values passed in are
coerced to open in the foreground.

BUG=491804

Review URL: https://codereview.chromium.org/1153193003

Cr-Commit-Position: refs/heads/master@{#331747}