Stop deleting the tab model metadata file immediately after loading.
Instead of immediately deleting the file on startup, hold onto it and
simply overwrite it when necessary.  This prevents cases where Chrome
hasn't managed to read in all of the TabState files before a crash,
which results in total annihilation of the tabs on the next load.

BUG=485217

Review URL: https://codereview.chromium.org/1156343002

Cr-Commit-Position: refs/heads/master@{#331832}