Change throbber (spinner) on tab strip to comply with material design
The previous model was to rotate a static image resource. We are replacing
it by a new model where the head revolves around the center while the tail
catches the head once in a while in an accelerated manner.

Instead of decoding throbber.svg file, we are calling paint_throbber.cc to
avoid file access. Caching was considered, but as you can see in
paint_throbber.cc, it would take quite a long time to get to the origin
because of its graphical nature, so caching would not be very efficient.

One visual caveat is that the current frame rate is limited to ~16 FPS
inside StripLayoutHelper, so the motion will stutter a bit when the user
is not taking any action. However, increasing FPS may have an adverse
effect on power consumption and performance, and some say it is hardly
noticeable, so I'll take care of it in a separate CL if many point it out.

BUG=489488

Review URL: https://codereview.chromium.org/1155583005

Cr-Commit-Position: refs/heads/master@{#332168}