Pull methods upwards from ChromeTabCreator.
Pulls methods from the Impl into the Interface so that
ChromeActivity can take in a generic TabCreator instead of
a ChromeTabCreator implementation.

This is prep work for making it so that document mode's TabDelegate is also a TabCreator, which will let it be
usable in webapp mode for handling window.open calls.

BUG=451453

Review URL: https://codereview.chromium.org/1163173002

Cr-Commit-Position: refs/heads/master@{#333094}