Fix theme color transition
The check at the beginning of onPrimaryColorChanged should have
checked whether the feature is enabled, but it used a state that
is dependent on the current color. Use isDOcumentMode instead.

Also changes the test file deflaking one test. Added a new test
to check transition to and from a non theme colored page.

BUG=497610

Review URL: https://codereview.chromium.org/1170813002

Cr-Commit-Position: refs/heads/master@{#333527}