Do not destroy incognito while tab is being created
If a close all incognito tabs operation is in progress while a tab is being created,
destroyIncognitoIfNecessary gets invoked as part of the close all tabs animation
finishing. This causes chrome to crash. Introducing a boolean isTabCreating and not
destroying incognito if it is true eliminates the crash.

BUG=496651

Review URL: https://codereview.chromium.org/1208403002

Cr-Commit-Position: refs/heads/master@{#336481}