Make urlbar in custom tabs not focusable
The urlbar in custom tabs is disabled and is not editable. However we
call urlbar.requestFocus in several cases, which causes a framework
crash. This CL added a flag mShouldHaveFocus in UrlBar and in onDraw()
we can choose not to set the urlbar to be focusable.

BUG=496418

Review URL: https://codereview.chromium.org/1207523003

Cr-Commit-Position: refs/heads/master@{#336641}