Only set precache.last_time when precache ran.
Modify the prefetch run to pass a precacheStarted boolean from the C++ side to
the Java side. If precacheStarted, the precache.last_time pref is set and an
alarm is created for WAIT_UNTIL_NEXT_PRECACHE_MS later. Otherwise, the pref is
not updated, and an alarm is set for INTERACTIVE_STATE_POLLING_PERIOD_MS later.

This should reduce the possibility that a user has to wait an additional 4
hours for the prefetch run, because some precondition for prefetching wasn't
yet met.

BUG=309216

Review URL: https://codereview.chromium.org/1272443002

Cr-Commit-Position: refs/heads/master@{#342833}