Add bookmark search feature to enhanced bookmark
Frontend:
1. The UI for bookmark search is implemented as a custom substitutable
View, which on phone takes the entire screen and behave as if an activity; on
tablet replaces the content RecyclerView and shows on the right of the screen.
2. The view tansitions are done by ViewSwitchers, which are placed as the root
of eb_main.xml.
3. EnhancedBookmarkSearchRow, a subclass of EnhancedBookmarkBookmarkRow, was
created to show bookmark results.

Backend:
1. We used the same backend that powers omnibox bookmark query. It
traverses the entire bookmark model and gets prefix matching resutls. A maximum
limit of 500 bookmarks is chosen in case of too large query that blocks the UI
thread.
2. Search history is saved in one single Android SharedPreference. Android does
not have a built-in support for saving list of strings; consequently JSONArray
is chosen to help serializing and deserializing.

The majority of the code has been previously reviewed in
https://chrome-internal-review.googlesource.com/#/c/180705/ and
https://chrome-internal-review.googlesource.com/#/c/216387/

BUG=515617

Review URL: https://codereview.chromium.org/1270693004

Cr-Commit-Position: refs/heads/master@{#343105}