Suspend Webkit shared timer periodically and clear memory if tab goes to background
This patch saves power when there are multiple tabs open by intermittently
suspending the WebKit shared timer to pause layout activities. It also frees
up memory by clearing all caches and doing full GC in V8.

The patch also modifies a test which was checking for a redirect in a
background tab immediately. We introduce a delay in it to wait until the timer
has resumed and the redirect finally happens

Change-Id: I64eb6a18c9d6d23880db3ca0b52f4851d50d9cbb